better examples

+++ examples/desktop/1-minimal_example.cpp
@@ -25,8 +25,8 @@ int main() {
     for (int i = 0; i < 10; ++i) {
         auto frame = cameraProvider.grab(3000);
         if (frame) {
-            printf("VideoFrame %lld grabbed: width = %d, height = %d, bytes: %d, format: %s\n", frame->frameIndex, frame->width,
-                   frame->height, frame->sizeInBytes, ccap::pixelFormatToString(frame->pixelFormat).data());
+            printf("VideoFrame %d grabbed: width = %d, height = %d, bytes: %d, format: %s\n", (int)frame->frameIndex, frame->width,
+                   frame->height, (int)frame->sizeInBytes, ccap::pixelFormatToString(frame->pixelFormat).data());
         } else {
             std::cerr << "Failed to grab frame!" << std::endl;
             exit(-1);
